+++ quake3-rs/ioquake3/src/opus_1_2_1/src/opus_multistream_decoder.rs
@@ -450,7 +450,7 @@ unsafe extern "C" fn opus_multistream_decode_native(
         (&mut Fs as *mut crate::opus_types_h::opus_int32).offset(
             (&mut Fs as *mut crate::opus_types_h::opus_int32)
                 .offset_from(&mut Fs as *mut crate::opus_types_h::opus_int32)
-                as libc::c_long as isize,
+                as isize as isize,
         ),
     );
     frame_size = if frame_size < Fs / 25 as i32 * 3 as i32 {
